add `\start...` and `\stop...` syntax
Specifically, any time we a command that begins with `start`
(e.g. `\starttext`) or that begins with `stop` (e.g. `\stoptext`), we
highlight it as `keyword.control`.

There are a couple of rules that relate to specific commands of this
type, so we have to make this change in a couple of different places.